netlify/functions/aurore/requirements.txt
requests
beautifulsoup4
PyGithub
Jinja2
tweepy
google-generativeai
python-slugify
netlify-blob